Fixed issues with updating buffer size etc.
It was just an issue with the message that was being printed. Also, when
changing the buffer size, we need to change the transfer function on the
device side, so that's been updated.

I also turned the time-string timer into a general-purpose listener,
which also listens for whether the song has ended every 100 ms. Any
other listening code should be added to the songListener() function.

Lastly, I've been changing the sliders so that we can possibly make the
whole GUI black.