Major GUI update; QCustomDial added.
The GUI (except for the QCustomPlot) has a mostly consistent font
scheme, color scheme, alignment scheme, and placement scheme. To find
out about the colors, look at the "styleSheet" property of various
elements in qtcreator. Note that the properties of the entire window can
be configured by changing the styleSheet of the "centralWidget" (NOT the
styleSheet of the main window).

I've also added a custom dial that looks nicer and allows for
logarithmically-spaced frequencies.