Added high/low-shelving filters; GUI tweaks.
The filters seem to be working, and are reasonably fast since they use
__expf when necessary. I also made some tweaks to the GUI, including
setting "disabled" colors (for buttons / spin boxes that get disabled)
and clamping the y-axis range of the real-time transfer function plot.
The latter change means that the graphs can sometimes go off the plot,
but that's mostly a visual cue to the user (like in FL Studio).