"Save" button implemented.
The position looks kind of awkward right now, but I don't know
where is the best place to put it. It is disabled when the song
is not loaded/is playing, and only enabled when we have stopped
from playing a song.

In addition, README was modified for installation information.