Fixed some saving bugs; GUI tweaks.
There were some bugs related to when saving would be enabled/disabled,
and how much of the song would be saved (previously, the entire song was
being saved). I also changed the saving code to use SFML instead of the
wavFile struct (which I removed).

There's still an issue where the saved file's recorded length (in
seconds) is twice the actual length for stereo files. In other words,
the files stop playing at the correct time, but their length is
incorrect. This issue was there with both SFML saving and manual saving
with the wavFile struct, and doesn't seem to be resolvable by dividing
by the number of channels (the SoundBuffer *says* it recorded sound of
the right duration; there's just something wrong with saving).

I also updated the GUI a bit.